L2 agent RPC add bulk calls for device up and device down
update_device_up and update_device_down are the RPC calls
used by the L2 agent to notify the plugin that
a device is up or down. These calls accept only one device.
This patch introduces two new RPC calls, update_device_list_up and
update_device_list_down that accept a list of devices. These calls
will return a list of devices for which the operation failed.
The agent will be able to update several devices using one RPC call
and in a following patch to retry the operation only for the devices
that failed instead of performing a full resync.

This patch introduces the new calls only for the OVS agent.
A following patch will take care of the other L2 agents.

Partially-Implements: blueprint restructure-l2-agent

Change-Id: Idb4e81f4399cf66025ec7edfcb78801c15e493af