Add state_or_sec_group_updated flag in port_update
port_update notification is sent when an attribute
of a port is modified. The OVS agent uses
this notification to keep track of the ports that
need to be reprocessed. Anyway not all the changes
that can affect a port require the L2 to react.
Add a flag in port_update to indicate that the state,
the IP or the security group changed for a port.
The L2 agent can use this information to avoid processing
a port when it's not needed.

Change-Id: I0ce3527a2f218f3fda231f724b82a68e031a23be
Partially-Implements: blueprint restructure-l2-agent