Introduce get_devices_details_list_and_failed_devices
get_devices_details_list_and_failed_devices is a new RPC
call that returns a list of devices details and a list of
devices for which it was not possible to retrieve the
details. The OVS agent can use that list and instead of
performing a full sync, will retry the operation only for
the failed devices at next iteration. This will be
implemented in another patch.

DocImpact
Partially-Implements: blueprint restructure-l2-agent

Change-Id: I54d19756bf94ec86c911f412b26dd6ddaeb93db9