Add updated_attrs in port_update
port_update notification is sent when an attribute
of a port is modified. The OVS agent uses
that notification to keep track of the ports that
need to be reprocessed. Anyway not all the changes
that can affect a port require the port to be
reprocessed. Add updated_attrs in port_update to
keep track of which attribute was updated. The
L2 agent uses that list to determine if a port
needs to be reprocessed or if updating the firewall
is enough.

Change-Id: I0ce3527a2f218f3fda231f724b82a68e031a23be
Partially-Implements: blueprint restructure-l2-agent