Move db_list to OVSBridge
OVSBridge was inheriting db_list from BaseOVS, which was
returning the information of all the ports on the machine,
not only the ones belonging to the bridge. This is
counterintuitive.
The OVSNeutronAgent is using that method with the assumption
that ports are filtered by bridge.
This patch moves db_list to OVSBridge and take into
consideration only the ports beloning to the bridge.
ovs_lib methods that use db_list are refactored accordingly.

Change-Id: I2ce6d232744f48ba7fc0f824a7db32e3655bc2aa
Closes-bug: 1473199