sync_allocations throws DBDuplicatedEntry with postgres HA
When using Galera and two (or more) processes tries to modify
the same records we get a deadlock exception.
Postgres in HA behaves differently, the process that commits
first sees no error, the others get a DBDuplicatedEntry
exception. As already done in Galera case, the solution is to
retry. The transaction will be restarted, the process will get
fresh data and won't try to insert the duplicated entries
anymore.

Change-Id: I7e257f6f2d1db3103e74170b865b184033b495dd
Closes-bug: #1421626